refactor: add utils, types and constants under workspace creator folder
