+++ src/plugins/workspace/public/components/workspace_creator/workspace_form.tsx
@@ -30,12 +30,10 @@ import {
 } from '@elastic/eui';
 import { i18n } from '@osd/i18n';
 import {
-  App,
   AppNavLinkStatus,
   ApplicationStart,
   DEFAULT_APP_CATEGORIES,
   MANAGEMENT_WORKSPACE_ID,
-  WorkspacePermissionMode,
 } from '../../../../../core/public';
 import { useApplications } from '../../hooks';
 import {
@@ -50,13 +48,18 @@ import {
 } from '../utils/feature';
 import { WorkspaceBottomBar } from './workspace_bottom_bar';
 import { WorkspaceIconSelector } from './workspace_icon_selector';
+import { WorkspacePermissionSettingPanel } from './workspace_permission_setting_panel';
+import { featureMatchesConfig } from '../../utils';
+import { getErrorsCount, getUnsavedChangesCount, isValidWorkspacePermissionSetting } from './utils';
 import {
-  getPermissionModeId,
+  WorkspaceFeature,
+  WorkspaceFeatureGroup,
+  WorkspaceFormData,
+  WorkspaceFormErrors,
+  WorkspaceFormSubmitData,
   WorkspacePermissionSetting,
   WorkspacePermissionItemType,
-  WorkspacePermissionSettingPanel,
-} from './workspace_permission_setting_panel';
-import { featureMatchesConfig } from '../../utils';
+} from './types';
 
 enum WorkspaceFormTabs {
   NotSelected,
@@ -65,46 +68,9 @@ enum WorkspaceFormTabs {
   UsersAndPermissions,
 }
 
-interface WorkspaceFeature extends Pick<App, 'dependencies'> {
-  id: string;
-  name: string;
-}
-
-interface WorkspaceFeatureGroup {
-  name: string;
-  features: WorkspaceFeature[];
-}
-
-export interface WorkspaceFormSubmitData {
-  name: string;
-  description?: string;
-  features?: string[];
-  color?: string;
-  icon?: string;
-  defaultVISTheme?: string;
-  permissions: WorkspacePermissionSetting[];
-}
-
-export interface WorkspaceFormData extends WorkspaceFormSubmitData {
-  id: string;
-  reserved?: boolean;
-}
-
-type WorkspaceFormErrors = Omit<{ [key in keyof WorkspaceFormData]?: string }, 'permissions'> & {
-  permissions?: string[];
-};
-
 const isWorkspaceFeatureGroup = (
   featureOrGroup: WorkspaceFeature | WorkspaceFeatureGroup
 ): featureOrGroup is WorkspaceFeatureGroup => 'features' in featureOrGroup;
-
-const isValidWorkspacePermissionSetting = (
-  setting: Partial<WorkspacePermissionSetting>
-): setting is WorkspacePermissionSetting =>
-  !!setting.modes &&
-  setting.modes.length > 0 &&
-  ((setting.type === WorkspacePermissionItemType.User && !!setting.userId) ||
-    (setting.type === WorkspacePermissionItemType.Group && !!setting.group));
 
 const isDefaultCheckedFeatureId = (id: string) => {
   return DEFAULT_CHECKED_FEATURES_IDS.indexOf(id) > -1;
@@ -122,145 +88,6 @@ const isValidNameOrDescription = (input?: string) => {
   const regex = /^[0-9a-zA-Z()_\[\]\-\s]+$/;
   return regex.test(input);
 };
-
-const getErrorsCount = (formErrors: WorkspaceFormErrors) => {
-  let errorsCount = 0;
-  if (formErrors.name) {
-    errorsCount += 1;
-  }
-  if (formErrors.description) {
-    errorsCount += 1;
-  }
-  if (formErrors.permissions) {
-    errorsCount += formErrors.permissions.length;
-  }
-  return errorsCount;
-};
-
-// when editing, attributes could be undefined in workspace form
-type WorkspaceFormEditingData = Partial<
-  Omit<WorkspaceFormSubmitData, 'permissions'> & {
-    permissions: Array<Partial<WorkspacePermissionSetting>>;
-  }
->;
-
-type UserOrGroupPermissionEditingData = Array<
-  Partial<{ id: string; modes: WorkspacePermissionMode[] }>
->;
-
-const getUserAndGroupPermissions = (permissions: Array<Partial<WorkspacePermissionSetting>>) => {
-  const userPermissions: UserOrGroupPermissionEditingData = [];
-  const groupPermissions: UserOrGroupPermissionEditingData = [];
-  if (permissions) {
-    for (const permission of permissions) {
-      if (permission.type === WorkspacePermissionItemType.User) {
-        userPermissions.push({ id: permission.userId, modes: permission.modes });
-      }
-      if (permission.type === WorkspacePermissionItemType.Group) {
-        groupPermissions.push({ id: permission.group, modes: permission.modes });
-      }
-    }
-  }
-  return [userPermissions, groupPermissions];
-};
-
-const getUnsavedUserOrGroupPermissionChangesCount = (
-  initialPermissions: UserOrGroupPermissionEditingData,
-  currentPermissions: UserOrGroupPermissionEditingData
-) => {
-  // for user or group permissions, unsaved changes is the sum of
-  // # deleted permissions, # added permissions and # edited permissions
-  let addedPermissions = 0;
-  let editedPermissions = 0;
-  const initialPermissionMap = new Map<string, WorkspacePermissionMode[]>();
-  for (const permission of initialPermissions) {
-    if (permission.id) {
-      initialPermissionMap.set(permission.id, permission.modes ?? []);
-    }
-  }
-
-  for (const permission of currentPermissions) {
-    if (!permission.id) {
-      addedPermissions += 1; // added permissions
-    } else {
-      const permissionModes = initialPermissionMap.get(permission.id);
-      if (!permissionModes) {
-        addedPermissions += 1;
-      } else if (
-        getPermissionModeId(permissionModes) !== getPermissionModeId(permission.modes ?? [])
-      ) {
-        editedPermissions += 1; // added or edited permissions
-      }
-    }
-  }
-
-  // currentPermissions.length = initialPermissions.length + # added permissions - # deleted permissions
-  const deletedPermissions =
-    addedPermissions + initialPermissions.length - currentPermissions.length;
-
-  return addedPermissions + editedPermissions + deletedPermissions;
-};
-
-const getUnsavedChangesCount = (
-  initialFormData: WorkspaceFormEditingData,
-  currentFormData: WorkspaceFormEditingData
-) => {
-  let unsavedChangesCount = 0;
-  if (initialFormData.name !== currentFormData.name) {
-    unsavedChangesCount += 1;
-  }
-  // initial and current description could be undefined
-  const initialDescription = initialFormData.description ?? '';
-  const currentDescription = currentFormData.description ?? '';
-  if (initialDescription !== currentDescription) {
-    unsavedChangesCount += 1;
-  }
-  if (initialFormData.color !== currentFormData.color) {
-    unsavedChangesCount += 1;
-  }
-  if (initialFormData.icon !== currentFormData.icon) {
-    unsavedChangesCount += 1;
-  }
-  if (initialFormData.defaultVISTheme !== currentFormData.defaultVISTheme) {
-    unsavedChangesCount += 1;
-  }
-  const featureIntersectionCount = (
-    initialFormData.features?.filter((feature) => currentFormData.features?.includes(feature)) ?? []
-  ).length;
-  // for features, unsaved changes is the sum of # deleted features and # added features
-  unsavedChangesCount += (initialFormData.features?.length ?? 0) - featureIntersectionCount;
-  unsavedChangesCount += (currentFormData.features?.length ?? 0) - featureIntersectionCount;
-  // for permissions, unsaved changes is the sum of # unsaved user permissions and # unsaved group permissions
-  const [initialUserPermissions, initialGroupPermissions] = getUserAndGroupPermissions(
-    initialFormData.permissions ?? []
-  );
-  const [currentUserPermissions, currentGroupPermissions] = getUserAndGroupPermissions(
-    currentFormData.permissions ?? []
-  );
-  unsavedChangesCount += getUnsavedUserOrGroupPermissionChangesCount(
-    initialUserPermissions,
-    currentUserPermissions
-  );
-  unsavedChangesCount += getUnsavedUserOrGroupPermissionChangesCount(
-    initialGroupPermissions,
-    currentGroupPermissions
-  );
-  return unsavedChangesCount;
-};
-
-const isUserOrGroupPermissionSettingDuplicated = (
-  permissionSettings: Array<Partial<WorkspacePermissionSetting>>,
-  permissionSettingToCheck: WorkspacePermissionSetting
-) =>
-  permissionSettings.some(
-    (permissionSetting) =>
-      (permissionSettingToCheck.type === WorkspacePermissionItemType.User &&
-        permissionSetting.type === WorkspacePermissionItemType.User &&
-        permissionSettingToCheck.userId === permissionSetting.userId) ||
-      (permissionSettingToCheck.type === WorkspacePermissionItemType.Group &&
-        permissionSetting.type === WorkspacePermissionItemType.Group &&
-        permissionSettingToCheck.group === permissionSetting.group)
-  );
 
 const workspaceHtmlIdGenerator = htmlIdGenerator();
 
@@ -819,6 +646,7 @@ export const WorkspaceForm = ({
           <EuiTab
             onClick={handleTabFeatureClick}
             isSelected={selectedTab === WorkspaceFormTabs.FeatureVisibility}
+            data-test-subj="workspaceForm-tabSelection-featureVisibility"
           >
             <EuiText>{featureVisibilityTitle}</EuiText>
           </EuiTab>
@@ -827,6 +655,7 @@ export const WorkspaceForm = ({
           <EuiTab
             onClick={handleTabWorkspaceSettingsClick}
             isSelected={selectedTab === WorkspaceFormTabs.WorkspaceSettings}
+            data-test-subj="workspaceForm-tabSelection-workspaceSettings"
           >
             <EuiText>{workspaceSettingsTitle}</EuiText>
           </EuiTab>
@@ -835,6 +664,7 @@ export const WorkspaceForm = ({
           <EuiTab
             onClick={handleTabPermissionClick}
             isSelected={selectedTab === WorkspaceFormTabs.UsersAndPermissions}
+            data-test-subj="workspaceForm-tabSelection-usersAndPermissions"
           >
             <EuiText>{usersAndPermissionsTitle}</EuiText>
           </EuiTab>
